add policy to region/state/city

+++ app/controllers/api/v2/admin/regions_controller.rb
@@ -1,18 +1,19 @@
 class Api::V2::Admin::RegionsController < Api::V2::Admin::BaseController
 
   def index
-    regions = Region.all
-    paginate json: regions.order(:id), status: 200
+    regions = policy_scope [:admin, Region]
+    paginate json: regions, status: 200
   end
 
   def show
     region = Region.find(params[:id])
+    authorize [:admin, region]
     render json: region, status: 200
   end
 
   def create
     region = Region.new(region_params)
-
+    authorize [:admin, region]
     if region.save
       render json: region, status: 201
     else
@@ -22,6 +23,7 @@ class Api::V2::Admin::RegionsController < Api::V2::Admin::BaseController
 
   def update
     region = Region.find(params[:id])
+    authorize [:admin, region]
     if region.update(region_params)
       render json: region, status: 200
     else
@@ -31,6 +33,7 @@ class Api::V2::Admin::RegionsController < Api::V2::Admin::BaseController
 
   def destroy
     region = Region.find(params[:id])
+    authorize [:admin, region]
     region.destroy
     head 204
   end
@@ -38,6 +41,6 @@ class Api::V2::Admin::RegionsController < Api::V2::Admin::BaseController
   private
 
   def region_params
-    params.permit(:name)
+    params.permit(policy([:admin, Region]).permitted_attributes)
   end
 end
